Technical/Refactor truemail helpers
* Updated variadicValidationType(), tests
* Updated Validate(), IsValid(), tests
* Updated test helpers
